Refactorings & Cleanups
- Refactor yaml parsing into parse_yaml_with_spans
- Refactor Span and RuleLangError into rule_lang.py
- Fix alignment when the number of line numbers changes
- Remove xfail from tests (fixed cause of underlying failure)
