+++ semgrep/semgrep/semgrep_main.py
@@ -49,7 +49,7 @@ def validate_single_rule(config_id: str, rule: Dict[str, Any]) -> bool:
     if not rule_keys.issubset(YAML_ALL_VALID_RULE_KEYS):
         extra_keys = rule_keys - YAML_ALL_VALID_RULE_KEYS
         print_error(
-            f"{config_id} has invalid rule key {extra_keys} at rule id {rule_id_err_msg}, can only have: {YAML_DISPLAY_VALID_RULE_KEYS}"
+            f"{config_id} has invalid rule key {extra_keys} at rule id {rule_id_err_msg}, can only have: {sorted(list(YAML_DISPLAY_VALID_RULE_KEYS))}"
         )
         return False
     try:
