+++ semgrep/semgrep/semgrep_types.py
@@ -10,6 +10,8 @@ from typing import Set
 
 import attr
 
+from semgrep import rule_lang
+
 PatternId = NewType("PatternId", str)
 Operator = NewType("Operator", str)
 
@@ -42,17 +44,10 @@ OPERATOR_PATTERN_NAMES_MAP = {
     OPERATORS.REGEX: ["pattern-regex"],
 }
 
-START_LINE = "__line__"
-END_LINE = "__endline__"
-FILE = "__sourcefile__"
-RAW = "__raw__"
-
-SPAN_HINTS = {START_LINE, END_LINE, FILE, RAW}
-
 # These are the only valid top-level keys
 YAML_MUST_HAVE_KEYS = {"id", "message", "languages", "severity"}
 YAML_OPTIONAL_KEYS = {"metadata", "paths"}
-YAML_INTERNAL_KEYS = SPAN_HINTS
+YAML_INTERNAL_KEYS = rule_lang.SPAN_HINTS
 YAML_VALID_TOP_LEVEL_OPERATORS = {
     OPERATORS.AND,
     OPERATORS.AND_ALL,
@@ -79,47 +74,6 @@ class InvalidRuleSchema(BaseException):
     pass
 
 
-class Span(NamedTuple):
-    start_line: int  # 0 indexed
-    end_line: int  # 0 indexed
-    file: Optional[str]
-    raw: List[str]  # all lines in the file this span is in
-
-    context_start: int
-    context_end: int
-
-    @classmethod
-    def from_dict(cls, d: Dict[str, Any], before_context=0, after_context=0) -> Optional["Span"]:  # type: ignore
-        start_line = d.get(START_LINE)
-        end_line = d.get(END_LINE)
-        file = d.get(FILE)
-        raw: List[str] = d.get(RAW)  # type: ignore
-
-        if start_line is not None and end_line is not None:
-            context_start = max(start_line - before_context, 0)
-            context_end = min(end_line + after_context, len(raw))
-            return Span(
-                start_line,
-                end_line,
-                file,
-                raw,
-                context_start=context_start,
-                context_end=context_end,
-            )
-        else:
-            return None
-
-
-DUMMY_SPAN = Span(
-    start_line=0,
-    end_line=0,
-    file=None,
-    raw=["I am a dummy span."],
-    context_start=0,
-    context_end=0,
-)
-
-
 @attr.s(auto_attribs=True, frozen=True)
 class BooleanRuleExpression:
     operator: Operator
@@ -130,7 +84,7 @@ class BooleanRuleExpression:
     operand: Optional[str] = None
 
     # For tests, eg. don't force people to make spans
-    span: Optional[Span] = None
+    span: Optional[rule_lang.Span] = None
     raw: Optional[Dict[str, Any]] = None
 
     def __attrs_post_init__(self) -> None:
@@ -144,12 +98,10 @@ class BooleanRuleExpression:
                 )
         else:
             if self.children is not None:
-                from semgrep.error import SemgrepLangError
-
-                err = SemgrepLangError(
+                err = rule_lang.RuleLangError(
                     short_msg=f"{pattern_names_for_operator(self.operator)[0]} cannot have children",
                     long_msg=f"only {pattern_names_for_operators(OPERATORS_WITH_CHILDREN)} operators can have children",
-                    spans=[self.span or DUMMY_SPAN],
+                    spans=[self.span or rule_lang.DUMMY_SPAN],
                     level="error",
                 )
                 raise InvalidRuleSchema(err.emit())
@@ -160,12 +112,10 @@ class BooleanRuleExpression:
                 )
             else:
                 if type(self.operand) != str:
-                    from semgrep.error import SemgrepLangError
-
-                    err = SemgrepLangError(
+                    err = rule_lang.RuleLangError(
                         short_msg="invalid type",
                         long_msg=f"value for operator `{pattern_names_for_operator(self.operator)[0]}` must be a string, but was {type(self.operand).__name__}",
-                        spans=[self.span or DUMMY_SPAN],
+                        spans=[self.span or rule_lang.DUMMY_SPAN],
                         level="error",
                     )
                     raise InvalidRuleSchema(err.emit())
