+++ semgrep/tests/e2e/snapshots/test_rule_parser/test_rule_parser__failure__error_messages/bad1/error.txt
@@ -1,11 +1,11 @@
 error: pattern-inside cannot have children
   --> rules/syntax/bad1.yaml:4
-4 |       - pattern-inside:
-5 |         - pattern: $X == $X
-6 |         - pattern: $X != $X
-7 |         - patterns:
-8 |           - pattern-inside: |
-9 |               def __init__(...):
+4  |       - pattern-inside:
+5  |         - pattern: $X == $X
+6  |         - pattern: $X != $X
+7  |         - patterns:
+8  |           - pattern-inside: |
+9  |               def __init__(...):
 10 |                   ...
 11 |           - pattern: self.$X == self.$X
